requirements.txt
pygments>=2.17.0
pillow>=10.0.0
pathspec>=0.12.0
click>=8.1.0
